try downloading blob
